refactor: move cli interface out of jina module
