+++ MANIFEST.in
@@ -1,3 +1,4 @@
 recursive-include jina/resources *
 include extra-requirements.txt
 prune tests/
+prune **/tests/
