+++ jina/__init__.py
@@ -268,5 +268,4 @@ def set_nofile(nofile_atleast=4096):
     default_logger.debug(f'ulimit -n soft,hard: {soft} {hard}')
     return soft, hard
 
-
 set_nofile()
