+++ jina/checker.py
@@ -3,9 +3,9 @@ __license__ = "Apache-2.0"
 
 import os
 
-from .. import __jina_env__, import_classes
-from ..helper import colored, print_dep_tree_rst
-from ..logging import default_logger
+from . import __jina_env__, import_classes
+from .helper import colored, print_dep_tree_rst
+from .logging import default_logger
 
 if False:
     # fix type-hint complain for sphinx and flake
@@ -50,9 +50,9 @@ class NetworkChecker:
     """Check if a BasePod is running or not """
 
     def __init__(self, args: 'argparse.Namespace'):
-        from ..peapods.pea import send_ctrl_message
-        from ..proto import jina_pb2
-        from ..logging.profile import TimeContext
+        from jina.peapods.pea import send_ctrl_message
+        from jina.proto import jina_pb2
+        from jina.logging.profile import TimeContext
         from google.protobuf.json_format import MessageToJson
         import time
         ctrl_addr = 'tcp://%s:%d' % (args.host, args.port)
