+++ jina/peapods/container.py
@@ -24,7 +24,7 @@ class ContainerPea(BasePea):
         # the image arg should be ignored otherwise it keeps using ContainerPea in the container
         # basically all args in BasePea-docker arg group should be ignored.
         # this prevent setting containerPea twice
-        from ..main.parser import set_pea_parser
+        from ..parser import set_pea_parser
         non_defaults = get_non_defaults_args(self.args, set_pea_parser(),
                                              taboo={'uses', 'entrypoint', 'volumes', 'pull_latest'})
