+++ jina_cli/__init__.py
@@ -5,9 +5,9 @@ import sys
 
 
 def _get_run_args(print_args: bool = True):
-    from ..logging import default_logger
-    from .parser import get_main_parser
-    from ..helper import colored
+    from jina.logging import default_logger
+    from jina.parser import get_main_parser
+    from jina.helper import colored
 
     parser = get_main_parser()
     if len(sys.argv) > 1:
