+++ jina_cli/api.py
@@ -4,14 +4,14 @@ __license__ = "Apache-2.0"
 
 def pod(args):
     """Start a Pod"""
-    from ..peapods import Pod
+    from jina.peapods import Pod
     with Pod(args) as p:
         p.join()
 
 
 def pea(args):
     """Start a Pea"""
-    from ..peapods import Pea
+    from jina.peapods import Pea
     try:
         with Pea(args) as p:
             p.join()
@@ -21,43 +21,43 @@ def pea(args):
 
 def gateway(args):
     """Start a Gateway Pod"""
-    from ..peapods.pod import GatewayPod
+    from jina.peapods.pod import GatewayPod
     with GatewayPod(args) as fs:
         fs.join()
 
 
 def log(args):
     """Receive piped log output and beautify the log"""
-    from ..logging.pipe import PipeLogger
+    from jina.logging.pipe import PipeLogger
     PipeLogger(args).start()
 
 
 def check(args):
     """Check jina config, settings, imports, network etc"""
-    from .checker import ImportChecker
+    from jina.checker import ImportChecker
     ImportChecker(args)
 
 
 def ping(args):
-    from .checker import NetworkChecker
+    from jina.checker import NetworkChecker
     NetworkChecker(args)
 
 
 def client(args):
     """Start a client connects to the gateway"""
-    from ..clients.python import PyClient
+    from jina.clients.python import PyClient
     PyClient(args)
 
 
 def export_api(args):
     from .export import api_to_dict
-    from .. import __version__
-    from ..logging import default_logger
+    from jina import __version__
+    from jina.logging import default_logger
 
     if args.yaml_path:
         for yp in args.yaml_path:
             f_name = (yp % __version__) if '%s' in yp else yp
-            from ..helper import yaml
+            from jina.helper import yaml
             with open(f_name, 'w', encoding='utf8') as fp:
                 yaml.dump(api_to_dict(), fp)
             default_logger.info(f'API is exported to {f_name}')
@@ -72,13 +72,13 @@ def export_api(args):
 
 
 def hello_world(args):
-    from ..helloworld import hello_world
+    from jina.helloworld import hello_world
     hello_world(args)
 
 
 def flow(args):
     """Start a Flow from a YAML file or a docker image"""
-    from ..flow import Flow
+    from jina.flow import Flow
     if args.uses:
         f = Flow.load_config(args.uses)
         with f:
