+++ jina_cli/export.py
@@ -5,9 +5,9 @@ import os
 
 
 def api_to_dict():
-    from ..enums import BetterEnum
-    from .. import __version__
-    from .parser import get_main_parser
+    from jina.enums import BetterEnum
+    from jina import __version__
+    from jina.parser import get_main_parser
 
     from argparse import _StoreAction, _StoreTrueAction
     port_attr = ('help', 'choices', 'default', 'required', 'option_strings', 'dest')
