+++ tests/integration/hub_usage/test_hub_usage.py
@@ -7,7 +7,7 @@ from jina.docker.hubio import HubIO
 from jina.excepts import PeaFailToStart
 from jina.executors import BaseExecutor
 from jina.flow import Flow
-from jina.main.parser import set_pod_parser, set_hub_build_parser
+from jina.parser import set_pod_parser, set_hub_build_parser
 from jina.peapods import Pod
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
