+++ tests/unit/docker/test_hub_build.py
@@ -4,7 +4,7 @@ import pytest
 
 from jina.docker.hubio import HubIO
 from jina.helper import yaml
-from jina.main.parser import set_hub_build_parser, set_hub_pushpull_parser
+from jina.parser import set_hub_build_parser, set_hub_pushpull_parser
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
