+++ tests/unit/flow/test_flow.py
@@ -6,8 +6,8 @@ import requests
 from jina import JINA_GLOBAL
 from jina.enums import FlowOptimizeLevel, SocketType
 from jina.flow import Flow
-from jina.main.checker import NetworkChecker
-from jina.main.parser import set_pea_parser, set_ping_parser, set_flow_parser, set_pod_parser
+from jina.checker import NetworkChecker
+from jina.parser import set_pea_parser, set_ping_parser, set_flow_parser, set_pod_parser
 from jina.peapods.pea import BasePea
 from jina.peapods.pod import BasePod
 from jina.proto.jina_pb2 import Document
