+++ tests/unit/test_driver_yaml.py
@@ -6,7 +6,7 @@ from jina.drivers.control import ControlReqDriver
 from jina.drivers.search import KVSearchDriver
 from jina.executors import BaseExecutor
 from jina.helper import yaml
-from jina.main.parser import set_pod_parser
+from jina.parser import set_pod_parser
 from jina.peapods import Pod
 from pkg_resources import resource_filename
 from tests import JinaTestCase
