+++ tests/unit/test_index.py
@@ -8,7 +8,7 @@ import numpy as np
 from jina.enums import FlowOptimizeLevel
 from jina.executors.indexers.vector import NumpyIndexer
 from jina.flow import Flow
-from jina.main.parser import set_flow_parser
+from jina.parser import set_flow_parser
 from jina.proto import jina_pb2
 from tests import JinaTestCase, random_docs
