+++ tests/unit/test_index_remote.py
@@ -7,7 +7,7 @@ import numpy as np
 from jina.enums import FlowOptimizeLevel
 from jina.executors.indexers.vector import NumpyIndexer
 from jina.flow import Flow
-from jina.main.parser import set_gateway_parser
+from jina.parser import set_gateway_parser
 from jina.peapods.pod import GatewayPod
 from tests import JinaTestCase, random_docs
